[lib] Fixed empty submatrix check
- Fix: Eliminated unsigned integer overflow in empty submatrix check.
  - The check used to fail when the difference between a pair of indexes is less than -1 i.e (x - y < -1).
  - Failures of the check resulted in allocation of very large blocks of memory.

+++ src/mat47/matrix.c
@@ -248,12 +248,12 @@ mat47_get_submat
         || check_col(m, right)
     ) return NULL;
 
-    if ((n_rows = bottom - top + 1) < 1) {
+    if ((n_rows = (long)bottom - top + 1) < 1) {
         mat47_errno = MAT47_ERR_ZERO_SIZE;
         error(": top=%u, bottom=%u", top, bottom);
         return NULL;
     }
-    if ((n_cols = right - left + 1) < 1) {
+    if ((n_cols = (long)right - left + 1) < 1) {
         mat47_errno = MAT47_ERR_ZERO_SIZE;
         error(": left=%u, right=%u", left, right);
         return NULL;
